[CRUD][LISTINGS] Moved ListingInstance to main dir and changed listing rep to array instead of object
